requirements.txt
requests
beautifulsoup4
confluent_kafka
kafka-python
psycopg2-binary
polars
pandas
python-dateutil
pyarrow
numpy
pymongo
openai
chonkie
tokenizers
python-dotenv
ollama
streamlit
PyPDF2
python-multipart